feat: Use upstream OverlapProvider and MergeProvider
Replace hand-crafted SQL generation for overlap with the upstream
OverlapProvider from datafusion-bio-function-ranges, and replace
the ~160-line Python window-function merge implementation with the
upstream MergeProvider routed through the Rust backend.

This gains file-path and LazyFrame streaming support for merge
(previously DataFrame-only) and removes ~260 lines of code.

+++ polars_bio/polars_ext.py
@@ -94,7 +94,7 @@ class PolarsRangesOperations:
 
     def merge(
         self,
-        min_dist: float = 0,
+        min_dist: int = 0,
         cols: Union[list[str], None] = None,
     ) -> pl.LazyFrame:
         """
